Move entity to model package
