+++ movieservice/src/main/java/com/drbsimon/movieservice/controller/MovieController.java
@@ -1,15 +1,10 @@
 package com.drbsimon.movieservice.controller;
 
-import com.drbsimon.movieservice.entity.Movie;
+import com.drbsimon.movieservice.model.Movie;
 import com.drbsimon.movieservice.model.MovieListWrapper;
 import com.drbsimon.movieservice.repository.MovieManager;
-import com.drbsimon.movieservice.repository.MovieRepository;
 import lombok.RequiredArgsConstructor;
-import org.slf4j.Logger;
 import org.springframework.web.bind.annotation.*;
-
-import java.security.AlgorithmConstraints;
-import java.util.List;
 
 //@CrossOrigin
 @RestController
