+++ movieservice/src/main/java/com/drbsimon/movieservice/repository/MovieRepository.java
@@ -1,6 +1,6 @@
 package com.drbsimon.movieservice.repository;
 
-import com.drbsimon.movieservice.entity.Movie;
+import com.drbsimon.movieservice.model.Movie;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
